fix: Support field names starting with a capital letter. (#39)
* fix: Support field names starting with a capital letter.

In GCE discovery API there are a few malformed field names starting with a capital letter, like `IPAddress`.

If converted to proto file as `ip_address` or `i_p_address` then in `JSON` it is serialized as `ipAddress` or `iPAddress` respectively, which would break `GlobalForwardingRulesClient.insert()` call. To be properly serialized into json as `IPAddress` in proto the field must be converted as `I_p_address`. See internal bug from a customer for more details.