feat: Introduce `--enums_as_strings` command line arg and feature (#59)
If `--enums_as_strings=true`, all enum fields except those are in LRO response message (normally called `Operation`) will be generate as `string` types.

By default `--enums_as_strings=false`

The enums definitions are still emitted regardless of the flag, but are not used as the actual field type. They still serve a purpose as a list of valid values for the string field and the enum is referenced from the corresponding field comment/documentation.