feat: Introduce proto merger
Proto merger merges elements from the previous version of the same proto file to avoid breaking changes.
There is no goal to have fully-featured merger, as it of itself would be a very difficult task. Instead it concentrates on merging the changes we already have to deal with, plus providing mechanism to easily add any additional merging funcitonality we may need in the future (hopefully none). Currently supported breaking changes merges:
- merging a missing field from old proto file in a new proto file
- merging a missing enum value from old proto to new
- merging mismatching service method google.api.method_signature option values

This PR also changes how conflicts in enum values are resolved on generation step. Previously, if there was at least one conflict, all the enums in the message were removed and corresponding fields were converted to `string` type automatically. Now, in case of the conflict the converter would simply remove the conflicting enums from all of the enums.