fix(temporary): reorder imports to work around protobuf.js problem (#108)
Apparently,  `protobuf.js` is not currently importing the `Any` type when `any.proto` is last in the import list, but the import does work if `any.proto` is higher in the import list. Pending a fix for this in `protobuf.js`, this change imports `any.proto` earlier to avoid triggering the problem. When `protobuf.js` is fixed, we can revert the change to maintain alphabetical ordering of imports.