fix: handle schema "format" google.protobuf.{Struct,Value,ListValue}
The schema "type" corresponding to each of these formats is as seen in
the baseline file `compute.any-format.proto.baseline`.

Googlers, also see cl/683528084

NOTE: that we continue to not handle `google.protobuf.Any`, whether it
arises from a format specification with that value or from a `"type":
"any"` with no "format" specified.