fix: handle type:object with format:google.protobuf.Any (#138)
We were not handling the Discovery equivalent of proto `repeated Any`, which shows up in Discovery docs as an `array` of `object` with `format:google.protobuf.Any`. This fixes that, and adds a test.

Note that because we only allow `google.protobuf.Any` when it's nested inside `....error.details`, the test case for this has more structure than other test cases for the `format` field.