fix: avoid request name conflicts: Discovery schema vs proto message
In the generated protos, the RPC request message is a wrapper around
the message corresponding to the request schema in the Discovery
file. The name of the proto request message was previously derived
purely from thes service and RPC name, with no consideration of
messages generated from explicit Discovery schemas.

This causes a problem when the name so generated exactly matches the
name of a Discovery schema. The name conflict leads to a
self-referential message that has the expected structure of the proto
request message, but is wrapping a field of the same type as itself.

This change fixes that by preventing such name conflicts and using a
different name for the proto RPC request message where such a conflict
would occur.