fix: avoid request name conflicts: Discovery schema vs proto message (#140)
In the generated protos, the RPC request message is a wrapper around the message corresponding to the request schema in the Discovery file. The name of the proto request wrapper message was previously derived purely from the service and RPC names, with no consideration of messages generated from explicit Discovery schemas.

This causes a problem when the proto request wrapper name so generated exactly matches the name of an explicit Discovery schema. The name conflict leads to a self-referential message that has the expected structure of the proto request wrapper message, but is wrapping a field of the same type as itself.

This change fixes that by preventing such name conflicts and using a different name for the proto RPC request wrapper message where such a conflict would occur.

Fixes [b/385211170](http://b/385211170)